[utilities] fix navbar,log,announcement
