--- ultilities/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-
-# Will not using Django Admin
