--- ultilities/static/js/popMessage.js
@@ -1,3 +1,0 @@
-function popMessage() {
-	
-}
