--- ultilities/templates/ultilities_make_announcement.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>iCovid - Make announcement</title>
-    <link rel="stylesheet" href="{% static 'https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css' %}" />
-    <link rel="stylesheet" type="text/javascript"  href="{% static 'js/includeHTML.js' %}" />
-    <link rel="stylesheet" type="text/html" href="{% static 'html/ultilities_navbar_admin.html' %}" />
-  </head>
-  <body>
-    <div include-html="ultilities_navbar_admin.html"></div>
-    <script>includeHTML()</script>
-
-    <form action="" method="POST">
-      {% csrf_token %}
-      <div class="form-group">
-        <label for="id_title">Title</label>
-        <input id="id_title" type="text" name="title" placeholder="Title" required="true"/>
-      </div>
-      <div class="form-group">
-        <label for="id_message">Message</label>
-        <textarea id="id_message" class="form-control" name="message"  rows="3" placeholder="Message" required="true"></textarea>
-      </div>
-      <br>
-      <button type="submit" class="btn btn-primary" style="background: #c42d4a;">Post</button>
-    </form>
-  </body>
-</html>
