--- ultilities/templates/ultilities_navbar_admin.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="stylesheet" type="text/css" href="{% static 'css/ultilities_styles.css' %}" />
-  </head>
-  <body>
-    <!-- Navbar Admin -->
-    <header id="ultilities">
-      <a id="ultilities" href="#home" style="margin-right: auto;"><img class="logo" src="{% static 'images/icovid-logo.png' %}" alt="logo" /></a>
-      <div class="drop" id="ultilities">
-        <button id="ultilities" class="dbutton">Admin</button>
-        <div class="dcontent" id="ultilities">
-          <a id="ultilities" href="#article"                        >Post article</a>
-          <a id="ultilities" href="/administrator/make-announcement">Make announcement</a>
-          <a id="ultilities" href="/administrator/see-log"          >See log</a>
-        </div>
-      </div> 
-      <nav> 
-        <ul class="navibar" id="ultilities">
-          <li id="ultilities"><a id="ultilities" href="#home"     >Home</a></li>  
-          <li id="ultilities"><a id="ultilities" href="#news"     >News</a></li>
-          <li id="ultilities"><a id="ultilities" href="#forum"    >Forum</a></li>
-          <li id="ultilities"><a id="ultilities" href="#tracking" >Tracking</a></li>
-          <li id="ultilities"><a id="ultilities" href="#home" onclick="" >Log out</a></li>  
-        </ul>
-      </nav>
-    </header>
-    <!--end navbar-->
-  </body>
-</html>
