--- ultilities/templates/ultilities_navbar_guest.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="stylesheet" type="text/css" href="{% static 'css/ultilities_styles.css' %}" />
-  </head>
-  <body>
-  	<!-- Navbar Guest -->
-    <header id="ultilities">
-      <a id="ultilities" href="#home" style="margin-right: auto;"><img class="logo" src="{% static 'images/icovid-logo.png' %}" alt="logo" /></a>
-      <nav> 
-        <ul class="navibar" id="ultilities">
-          <li id="ultilities"><a id="ultilities" href="#home"     >Home</a></li>  
-          <li id="ultilities"><a id="ultilities" href="#news"     >News</a></li>
-          <li id="ultilities"><a id="ultilities" href="#forum"    >Forum</a></li>
-          <li id="ultilities"><a id="ultilities" href="#tracking" >Tracking</a></li>
-          <li id="ultilities"><a id="ultilities" href="#auth"     >Log in</a></li>  
-        </ul>
-      </nav>
-      <a id="ultilities" href="#auth"><button id="ultilities">Sign up</button></a>
-    </header>
-    <!--end navbar--> 
-  </body>
-</html>
