--- ultilities/templates/ultilities_see_log.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>iCovid - See Log</title>
-    <link rel="stylesheet" type="text/javascript"  href="{% static 'js/includeHTML.js' %}" />
-    <link rel="stylesheet" type="text/html" href="{% static 'html/ultilities_navbar_admin.html' %}" />
-    <link rel="stylesheet" type="text/javascript"  href="{% static 'js/showLog.js' %}" />
-  </head>
-  <body>
-    <div include-html="ultilities_navbar_admin.html"></div>
-    <script>includeHTML()</script>
-
-    <div id="web-log"></div>
-    <script type="text/javascript" src="showLog.js"></script>
-  </body>
-</html>
