+++ utilities/templates/utilities_make_announcement.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+{% load static %}
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+    <title>iCovid - Make announcement</title>
+
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" />
+    <link rel="stylesheet" type="text/css" href="{% static 'css/utilities_styles.css' %}" />
+
+    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
+  </head>
+  <body>
+    {% include 'utilities_navbar_admin.html' %} 
+    
+    <!-- Make Announcement --> 
+    <section class="container-fluid">
+      <section class="row justify-content-center">
+        <section class="col-12 col-sm-6 col-md-3">
+          <form action="" method="POST" id="announcement" class="form-container">
+              {% csrf_token %}
+              <h4 class="text-center font-weight-bold">Make Announcement</h4>
+              <br>
+              <div class="form-group">
+                <input id="id_title" class="form-control" width="75%" type="text" name="title" required="true" placeholder="Title" />
+              </div>
+              <div class="form-group">
+                <textarea id="id_message" class="form-control" form="announcement" rows="4" cols="100" name="message" required="true" placeholder="Message"></textarea> 
+              </div>
+              <br>              
+              <button type="submit" class="btn btn-primary btn-block">Post</button>              
+          </form>
+        </section>
+      </section>
+    </section>
+    <!-- end section -->
+  </body>
+</html>
