+++ utilities/templates/utilities_modal_announcement.html
@@ -1,0 +1,40 @@
+<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" />
+
+<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
+    
+<!-- From utilities - add code in login method views.py authentication module 
+
+    from utilities.models import Announcement
+    announcement = Announcement.objects.last()
+    # include this in variable response => 'announcement': announcement
+
+-->
+
+<!--  From utilities - Attribute button for trigger popup
+    <button class="btn btn-primary" data-target="#announcementModal" data-toggle="modal">Login</button>
+-->
+
+<!-- From utilities - Modal Popup Announcement [add code in link '/login'] --> 
+<div class="modal fade" id="announcementModal" data-keyboard="false" data-backdrop="static" tabindex="-1">
+  <div class="modal-dialog">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h4 class="modal-title">Announcement</h4>
+        <button class="close" data-dismiss="modal">&times;</button>
+      </div>
+      <div class="modal-body">
+        <h5 style="text-align: center;">Title</h5>
+        <p style="text-align: center;">{{ announcement.title }}</p>
+        <br>
+        <h5 style="text-align: center;">Message</h5>
+        <p style="text-align: center;">{{ announcement.message }}</p>
+      </div>
+      <div class="modal-footer">
+        <button class="btn btn-info" data-dismiss="modal">Ok</button>
+      </div>
+    </div>
+  </div>
+</div>
+<!-- end modal -->
