+++ utilities/templates/utilities_navbar_admin.html
@@ -1,0 +1,23 @@
+<!-- Navbar Admin -->
+{% load static %}
+<header id="utilities">
+  <a id="utilities" href="#home" style="margin-right: auto;"><img class="logo" src="{% static 'images/icovid-logo.png' %}" alt="logo" /></a>
+  <div class="drop" id="utilities">
+    <button id="utilities" class="dbutton">Admin</button>
+    <div class="dcontent" id="utilities">
+      <a id="utilities" href="#article"                        >Post article</a>
+      <a id="utilities" href="/administrator/make-announcement">Make announcement</a>
+      <a id="utilities" href="/administrator/"                 >See log</a>
+    </div>
+  </div> 
+  <nav> 
+    <ul class="navibar" id="utilities">
+      <li id="utilities"><a id="utilities" href="#home"     >Home</a></li>  
+      <li id="utilities"><a id="utilities" href="#news"     >News</a></li>
+      <li id="utilities"><a id="utilities" href="#forum"    >Forum</a></li>
+      <li id="utilities"><a id="utilities" href="#tracking" >Tracking</a></li>
+      <li id="utilities"><a id="utilities" href="#home" onclick="" >Log out</a></li>  
+    </ul>
+  </nav>
+</header>
+<!--end navbar-->
