+++ utilities/templates/utilities_navbar_guest.html
@@ -1,0 +1,16 @@
+<!-- Navbar Guest -->
+{% load static %}
+<header id="utilities">
+  <a id="utilities" href="#home" style="margin-right: auto;"><img class="logo" src="{% static 'images/icovid-logo.png' %}" alt="logo" /></a>
+  <nav> 
+    <ul class="navibar" id="utilities">
+      <li id="utilities"><a id="utilities" href="#home"     >Home</a></li>  
+      <li id="utilities"><a id="utilities" href="#news"     >News</a></li>
+      <li id="utilities"><a id="utilities" href="#forum"    >Forum</a></li>
+      <li id="utilities"><a id="utilities" href="#tracking" >Tracking</a></li>
+      <li id="utilities"><a id="utilities" href="#auth"     >Log in</a></li>  
+    </ul>
+  </nav>
+  <a id="utilities" href="#auth"><button id="utilities">Sign up</button></a>
+</header>
+<!--end navbar-->
