+++ utilities/templates/utilities_navbar_user.html
@@ -1,0 +1,16 @@
+<!-- Navbar User -->
+{% load static %}
+<header id="utilities">
+  <a id="utilities" href="#home" style="margin-right: auto;"><img class="logo" src="{% static 'images/icovid-logo.png' %}" alt="logo" /></a>
+  <nav> 
+    <ul class="navibar" id="utilities">
+      <li id="utilities"><a id="utilities" href="#home"     >Home</a></li>  
+      <li id="utilities"><a id="utilities" href="#news"     >News</a></li>
+      <li id="utilities"><a id="utilities" href="#forum"    >Forum</a></li>
+      <li id="utilities"><a id="utilities" href="#tracking" >Tracking</a></li>
+      <li id="utilities"><a id="utilities" href="#profile"  >Profile</a></li>
+      <li id="utilities"><a id="utilities" href="#home" onclick="" >Log out</a></li>  
+    </ul>
+  </nav>
+</header>
+<!--end navbar-->
