+++ utilities/urls.py
@@ -3,5 +3,5 @@ from .views import make_announcement, see_log
 
 urlpatterns = [
     path('make-announcement', make_announcement, name='make_announcement'),
-    path('see-log', see_log, name='see_log'),
+    path('', see_log, name='see_log'),
 ]
